refactor: 관찰일지 결과 페이지를 서버 컴포넌트로 전환 및 ObservationBoundary 분리 (#74)
